make model and schema file different
